Implements basic UI with PySide6
Initializes the user interface using PySide6, replacing the command-line interface with a graphical one.

Adds a basic backend class to handle application logic. The backend is exposed to the QML interface.

Removes the command-line argument parsing and network-related code from `main.py`.
The network code is moved to tests for now.

Adds `PySide6` to the `requirements.txt`